Store luvi builtins in "luvi" module instead of on _G
Fixes #1
